% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.keys.R
\name{import.key}
\alias{import.key}
\title{import.key}
\usage{
import.key(path.root = NULL, str.api.name = "all")
}
\arguments{
\item{path.root}{Dropbox root path (default = ~/Dropbox/pkg.data)}

\item{str.api.name}{api key source (defualt = all, example = 'google')}
}
\description{
Imports api.key from local (dropbox) file
}
\keyword{api}
\keyword{key}
\keyword{load}
